% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad-data.R
\name{bad_data}
\alias{bad_data}
\title{Bingo squares for Bad Data}
\usage{
bad_data()
}
\value{
character vector of bingo square text
}
\description{
Inspired by the \href{https://github.com/Quartz/bad-data-guide}{Quartz Guide
to Bad Data}. Identifies bad data smells that should make you extremely
vigilant in your cleaning and sanity checks.
}
\examples{
bad_data()
}

