% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_games.R
\name{check_card}
\alias{check_card}
\title{Check a card for bingo}
\usage{
check_card(card, drawn)
}
\arguments{
\item{card}{A 5  x 5 matrix representing a bingo card}

\item{drawn}{A vector of spaces that have been drawn.  Should only contain
integers}
}
\value{
Logical, indicating if the card made a bingo
}
\description{
Check a card for bingo
}
\examples{
#none
}
