% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equal_cards.R
\name{check_equal_cards}
\alias{check_equal_cards}
\title{Checking equal cards}
\usage{
check_equal_cards(cards)
}
\arguments{
\item{cards}{list containing the generated cards}
}
\value{
none, only message
}
\description{
Checking equal cards
}
\examples{
cards <- bingo(10)
check_equal_cards(cards)
}
\author{
Marta Sestelo
}
