% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_games.R
\name{draw_spaces}
\alias{draw_spaces}
\title{Draw the order to call the spaces for a game}
\usage{
draw_spaces(cards)
}
\arguments{
\item{cards}{}
}
\value{
The order to call the spaces
}
\description{
Draw the order to call the spaces for a game
}
\examples{
#none
}
