% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{get_topic}
\alias{get_topic}
\title{Get the words of a bingo card topic}
\usage{
get_topic(topic)
}
\value{
A character vector containing all the possible words of a bingo card
topic
}
\description{
Get the words of a bingo card topic
}
\examples{
get_topic("open-data")
}

