% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{get_topics}
\alias{get_topics}
\title{Get all bingo card topics}
\usage{
get_topics()
}
\value{
A character vector containing the names of all the bingo card topics
}
\description{
Get all bingo card topics
}
\examples{
get_topics()
}
