% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch the Shiny app that generates bingo cards}
\usage{
launch(format = c("pdf", "html"))
}
\arguments{
\item{format}{Which format to generate the cards (downloadable PDF or
orintable HTML)}
}
\description{
Launch the Shiny app that generates bingo cards
}

