% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cards.R
\name{make_card}
\alias{make_card}
\title{Simulate a random bingo card}
\usage{
make_card(free_space = TRUE)
}
\arguments{
\item{free_space}{A logical value, whether or not to include a free space}
}
\value{
A dataframe with 5 rows and 5 columns to represent a bingo card
}
\description{
Simulate a random bingo card
}
\examples{
make_card()
}
