% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cards.R
\name{make_n_cards}
\alias{make_n_cards}
\title{Generate n bingo cards}
\usage{
make_n_cards(n = 8, free_space = TRUE)
}
\arguments{
\item{n}{A natural number, the number of bingo cards desired}
}
\value{
A list of length \code{n} bingo cards.
}
\description{
Generate n bingo cards
}
\examples{
#none
}
