% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-data.R
\name{open_data}
\alias{open_data}
\title{Bingo squares for "open" data and spreadsheets}
\usage{
open_data()
}
\value{
character vector of bingo square text
}
\description{
The agony of other people's data.
}
\examples{
open_data()
}

