% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cards.R
\name{sample_col}
\alias{sample_col}
\title{Sample column values}
\usage{
sample_col(off_set = 0)
}
\arguments{
\item{off_set}{A natural number, the lower bound for the column's values}
}
\description{
Sample column values
}
\examples{
#none
}
