% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{superbowl_50_2016}
\alias{superbowl_50_2016}
\title{Bingo squares for SuperBowl 50 (2016)}
\usage{
superbowl_50_2016()
}
\value{
character vector of bingo square text
}
\description{
Denver Broncos vs. Carolina Panthers
}
\examples{
superbowl_50_2016()
}

