% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\docType{data}
\name{topics}
\alias{topics}
\title{Environment containing all the word lists of all topics}
\format{An object of class \code{environment} of length 6.}
\usage{
topics
}
\description{
Environment containing all the word lists of all topics
}
\keyword{datasets}

