% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bingo}
\alias{bingo}
\title{Generate bingo cards}
\usage{
bingo(n_cards = 1, bs = NULL, n = 5)
}
\arguments{
\item{n_cards}{number of cards}

\item{bs}{text for the bingo squares}

\item{n}{number of rows/columns for one card}
}
\value{
a \code{bingo} object, which is really just a character matrix
}
\description{
Generate bingo cards
}
\examples{
bingo()
}

