% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bingo.R
\name{bingo}
\alias{bingo}
\title{Generate n bingo cards}
\usage{
bingo(ncards, seed = NULL)
}
\arguments{
\item{ncards}{number of cards}
}
\value{
a list with the generated cards
}
\description{
Generate n bingo cards
}
\examples{
cards <- bingo(10)
}
\author{
Marta Sestelo and Nora M. Villanueva
}
