% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{plot.bingo}
\alias{plot.bingo}
\title{Plot bingo cards}
\usage{
\method{plot}{bingo}(bc, dir = ".", fontsize = 14, pdf_base = "bingo-")
}
\arguments{
\item{bc}{a \code{\link{bingo}} object containing one or more bingo cards}

\item{dir}{directory where you want to write files}

\item{fontsize}{size of bingo square font}

\item{pdf_base}{base of the sequential filenames for the printable bingo card
files}
}
\value{
Vector containing the filenames of all the generated cards
  (invisibly)
}
\description{
Plot bingo cards
}
\note{
Does not actually plot the cards to the graphics device
}
\examples{
bc <- bingo()
plot(bc)
}

